Update index.html
The script tag for tailwind.config is not included correctly. It should be wrapped in a script tag with a type attribute.

You are using the tailwind.config script but have not included Tailwind CSS in your project. If you are not using Tailwind CSS, you can remove this script.

You have included jQuery and ScrollReveal scripts, but it seems you haven't used them in your code. If you don't need them, you can remove these scripts.

There is an issue with the font-awesome script inclusion. It should use the https protocol. Update the link as follows:


The <script src="https://cdn.tailwindcss.com"></script> script is unnecessary and can be removed if you're not using Tailwind CSS.

The <script async src="https://www.googletagmanager.com/gtag/js?id=G-80WSMNJP27"></script> and the related Google Analytics script are included, but they won't work unless you have a valid Google Analytics property ID. If you don't use Google Analytics, you can remove these scripts.

The href attribute for the <a> tags in the social media section is set to placeholder values. Make sure to replace them with your actual social media links.

The Twitter logo has a style attribute with an incorrect width property. Change width: full; to width: 100%;.

The scroll-reveal.js script is included, but the script file is not referenced correctly. Make sure to provide the correct path or remove it if not needed.

..